MDM V2.0 : Amélioration gestion du référentiel (int -> float + intégration)

+++ Mdm/src/fr/epita/sigl/miwa/application/PromotionForGC.java
@@ -5,11 +5,11 @@ import java.util.Date;
 public class PromotionForGC {
 	private String id;
 	private int nbMin;
-	private int rebate;
+	private float rebate;
 	private Date startDate;
 	private Date endDate;
 	
-	public PromotionForGC(String id, int nbMin, int pourcent, Date startDate, Date endDate) {
+	public PromotionForGC(String id, int nbMin, float pourcent, Date startDate, Date endDate) {
 		this.id = id;
 		this.nbMin = nbMin;
 		this.rebate = pourcent;
@@ -33,11 +33,11 @@ public class PromotionForGC {
 		this.nbMin = nbMin;
 	}
 
-	public int getRebate() {
+	public float getRebate() {
 		return rebate;
 	}
 
-	public void setRebate(int pourcent) {
+	public void setRebate(float pourcent) {
 		this.rebate = pourcent;
 	}
